feat: title page as separate dashboard item
