+++ src/components/Item/PrintTitlePageItem/styles/Item.module.css
@@ -1,0 +1,24 @@
+.container {
+    margin-left: 30px;
+}
+
+.title {
+    font-size: 20px;
+    font-weight: 500;
+    color: var(--colors-grey700);
+    cursor: default;
+    user-select: text;
+    margin: 0px 0px 0px 10px;
+}
+
+.description {
+    margin: 15px 0px 0px 10px;
+    font-size: 14px;
+    color: var(--colors-grey700);
+}
+
+.filters {
+    margin: 10px;
+    font-size: 14px;
+    color: var(--colors-grey700);
+}
